+++ jason/src/peer/media.rs
@@ -18,7 +18,6 @@ use web_sys::{
 
 use crate::{
     media::TrackConstraints,
-    peer::track::MutedState,
     utils::{JsCaused, JsError},
 };
 
@@ -28,6 +27,7 @@ use super::{
     stream_request::StreamRequest,
     track::MediaTrack,
 };
+use std::ops::Not;
 
 /// Errors that may occur in [`MediaConnections`] storage.
 #[derive(Debug, Display, JsCaused)]
@@ -389,6 +389,57 @@ impl MediaConnections {
             .or_else(|| {
                 inner.receivers.get(&id).and_then(|recv| recv.track.clone())
             })
+    }
+}
+
+/// Mute state of [`MediaTrack`].
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum MutedState {
+    /// [`MediaTrack`] is unmuted.
+    Unmuted,
+
+    /// [`MediaTrack`] should be unmuted, but awaits server permission.
+    Unmuting,
+
+    /// [`MediaTrack`] should be muted, but awaits server permission.
+    Muting,
+
+    /// [`MediaTrack`] is muted.
+    Muted,
+}
+
+impl MutedState {
+    /// Returns [`MutedState`] which should be set while transition to this
+    /// [`MutedState`].
+    pub fn proccessing_state(self) -> Self {
+        match self {
+            Self::Unmuted => Self::Unmuting,
+            Self::Muted => Self::Muting,
+            _ => self,
+        }
+    }
+}
+
+impl From<bool> for MutedState {
+    fn from(is_muted: bool) -> Self {
+        if is_muted {
+            Self::Muted
+        } else {
+            Self::Unmuted
+        }
+    }
+}
+
+impl Not for MutedState {
+    type Output = Self;
+
+    fn not(self) -> Self::Output {
+        match self {
+            Self::Muted => Self::Unmuted,
+            Self::Unmuted => Self::Muted,
+            Self::Unmuting => Self::Muting,
+            Self::Muting => Self::Unmuting,
+        }
     }
 }
